requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.2.1
drf-spectacular==0.24.2
email-validator==1.3.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.5
pyrsistent==0.19.2
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
sentry-sdk==1.11.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
